+++ CHANGELOG.md
@@ -16,3 +16,24 @@
 - cpp-class-generator.templates.header param
 - cpp-class-generator.templates.source param
 - "Generate C++ class files" menu option for folders
+
+## [v0.1.0b] - 25.07.2022
+
+### Added
+
+- FileFactory class
+- <%namespaceTab%> tag for namespace tabulation
+
+### Changed
+
+- cpp-class-generator.templates.copyright now contains the full comment
+- cpp-class-generator.tmeplates.header: removed comments from the beginning
+- cpp-class-generator.templates.single-header-extension: added .hh
+- cpp-class-generator.templates.header-extension: added .hh
+- cpp-class-generator.templates.source-extension: added .cc
+- code structure completely changed
+
+### Fixed 
+
+- [Issue#1](https://github.com/k4li-0x0/cpp-class-generator/issues/1)
+- [Issue#2](https://github.com/k4li-0x0/cpp-class-generator/issues/2)